+++ business/src/interest/model.rs
@@ -1,10 +1,11 @@
+use bigdecimal::BigDecimal;
 use uuid::Uuid;
 
 #[derive(Debug)]
 pub struct InterestConfiguration {
     pub id: Uuid,
-    pub start_range: i64,
-    pub end_range: i64,
-    pub fine: f64,
-    pub interest: f64,
+    pub start_range: i32,
+    pub end_range: i32,
+    pub fine: BigDecimal,
+    pub interest: BigDecimal,
 }
